Finished the functional programming challenges. Huzzah!
Got a bit stuck on "Sort an Array Alphabetically using the sort Method" since Chrome seems to have a different version of sort that treats characters differently(?) That was not so fun. Had to search the FreeCodeCamp forums to find that out. The "hint" had it wrong too - using the solution code failed in Chrome, in case it's unclear. That means perfectly valid JS code may not work properly in Chrome(ironically).

Currying is defintely a concept I haven't seen until I started learning Javascript.